Update to work with and require GregTechCE 1.17.0 (#158)
* Fix crash arising from Assembly Line energy-to-start datatype changing in GTCE 1.17.0
* Add/Update Assembly Line tooltips

+++ src/main/java/gregicadditions/machines/TileEntityFusionReactor.java
@@ -48,7 +48,7 @@ import net.minecraftforge.items.IItemHandlerModifiable;
 public class TileEntityFusionReactor extends RecipeMapMultiblockController {
 	private final int tier;
 	private EnergyContainerList inputEnergyContainers;
-	private int heat = 0; // defined in TileEntityFusionReactor but serialized in FusionRecipeLogic
+	private long heat = 0; // defined in TileEntityFusionReactor but serialized in FusionRecipeLogic
 
 	public TileEntityFusionReactor(ResourceLocation metaTileEntityId, int tier) {
 		super(metaTileEntityId, RecipeMaps.FUSION_RECIPES);
@@ -219,12 +219,12 @@ public class TileEntityFusionReactor extends RecipeMapMultiblockController {
 		@Override
 		protected Recipe findRecipe(long maxVoltage, IItemHandlerModifiable inputs, IMultipleTankHandler fluidInputs) {
 			Recipe recipe = super.findRecipe(maxVoltage, inputs, fluidInputs);
-			return (recipe != null && (long)recipe.getRecipePropertyStorage().getRecipePropertyValue(FusionEUToStartProperty.getInstance(), 0) <= energyContainer.getEnergyCapacity()) ? recipe : null;
+			return (recipe != null && recipe.getRecipePropertyStorage().getRecipePropertyValue(FusionEUToStartProperty.getInstance(), 0L) <= energyContainer.getEnergyCapacity()) ? recipe : null;
 		}
 
 		@Override
 		protected boolean setupAndConsumeRecipeInputs(Recipe recipe) {
-			int heatDiff = recipe.getRecipePropertyStorage().getRecipePropertyValue(FusionEUToStartProperty.getInstance(), 0) - heat;
+			long heatDiff = recipe.getRecipePropertyStorage().getRecipePropertyValue(FusionEUToStartProperty.getInstance(), 0L) - heat;
 			if (heatDiff <= 0) {
 				return super.setupAndConsumeRecipeInputs(recipe);
 			}
@@ -239,14 +239,14 @@ public class TileEntityFusionReactor extends RecipeMapMultiblockController {
 		@Override
 		public NBTTagCompound serializeNBT() {
 			NBTTagCompound tag = super.serializeNBT();
-			tag.setInteger("Heat", heat);
+			tag.setLong("Heat", heat);
 			return tag;
 		}
 
 		@Override
 		public void deserializeNBT(NBTTagCompound compound) {
 			super.deserializeNBT(compound);
-			heat = compound.getInteger("Heat");
+			heat = compound.getLong("Heat");
 		}
 	}
 }
